requirements.txt
# Core dependencies for video transcript extraction
yt-dlp>=2024.0.0
webvtt-py>=0.4.6
pandas>=2.0.0

# Optional: Semantic paragraph segmentation
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
scipy>=1.10.0
torch>=2.0.0
matplotlib>=3.5.0
